<Destroy new upgrade and mechanic> [New feature]
- upgrades for Destroy skill
- now destroy on level 1, 2 and 3 will have its neighboring tiles explode and hit characters inside it

+++ Assets/Scripts/Scriptable Object Scripts/Player Skill Data/SkillUpgradeData.cs
@@ -11,7 +11,7 @@ public class SkillUpgradeData
     [HideInInspector]
     public List<UPGRADE_BONUS> bonuses = new List<UPGRADE_BONUS>();
     [HideInInspector]
-    public List<float> additionalDamagePerLevel;
+    public List<int> additionalDamagePerLevel;
     [HideInInspector]
     public List<float> additionalPiercePerLevel;
     [HideInInspector]
